Fix devise error
Devise was throwing an error saying we couldn't have empty instance
variable for @user called in a app/views/devise/sessions/new.html.erb.
For some reason this view was changed along the commit history (I don't
know where) and that's what was causing the
error so we changed it back.